Implement login/signup pages
Create login and signup forms for the login signup pages.

The login and signup pages only offer a one-way connection with the rest
of the website (i.e. cannot get to login/signup from the main website)
for P1 because P1 requirements specify that all views are in a logged in
perspective.

Include simple js for toggling seeker/shelter tab and displaying the
error message. These toggles will lead to different landing pages when
the user clicks login.

Hardcode js to display an error message when the username is "abc".