Layout landing page and setup default cards
Implemented the general landing page with a blog section, a pet adoption
section and a shelter section. If not logged in, trying to click on
anything in pet adoption and shelter will point to the login page.

Set up default card style in global.css since they are used in multiple
pages.